requirements.txt
# CityBlitz – pinned requirements (Industrial Hardening)
# Prevents silent breakage from upstream releases. Update periodically and re-test.

streamlit>=1.32,<2.0
pandas>=2.1,<3.0
google-generativeai>=0.5,<1.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
altair>=5.2,<6.0

# Optional: encoding detection for non-UTF-8 CSVs (JanitorAgent)
chardet>=5.2,<6.0

# Render Web Service Dependencies
Flask>=2.3,<4.0
gunicorn==21.2.0
langchain>=0.1,<0.3
requests==2.31.0
python-dotenv==1.0.0
